style(Store): StoreCarousel 스타일 분리
