+++ src/components/StoreMain/StoreCarousel/index.js
@@ -1,10 +1,11 @@
-import Slider from 'react-slick';
-import 'slick-carousel/slick/slick.css';
-import 'slick-carousel/slick/slick-theme.css';
-
 import Image from 'next/image';
-import styled from 'styled-components';
+
 import { useRef } from 'react';
+import {
+  SliderButtonContainer,
+  SliderImage,
+  StyledSlider,
+} from './storeCarousel.style';
 
 const settings = {
   dots: true,
@@ -15,57 +16,6 @@ const settings = {
   autoplay: true,
   autoplaySpeed: 5000,
 };
-
-const StyledSlider = styled(Slider)`
-  .slick-dots {
-    bottom: 40px;
-
-    li {
-      width: 2px;
-      height: 2px;
-      margin: 0 7px;
-
-      button {
-        width: 2px;
-        height: 2px;
-      }
-      button:before {
-        width: 2px;
-        height: 2px;
-        color: white;
-      }
-    }
-  }
-`;
-
-const SliderImage = styled.div`
-  position: relative;
-  width: 100vw;
-  min-height: 400px;
-
-  img {
-    width: 100%;
-    height: 40vh;
-    object-fit: cover;
-    object-position: center center;
-  }
-`;
-
-const SliderButtonContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  button {
-    background-color: transparent;
-    border: 1px solid #ededed;
-    height: 40px;
-    padding: 0 10px;
-    cursor: pointer;
-  }
-  button:hover {
-    background-color: #ededed;
-    transition: 0.5s;
-  }
-`;
 
 export default function Carousel() {
   const customSliderRef = useRef();
