+++ src/pages/store/index.js
@@ -1,9 +1,13 @@
 import StoreMain from 'components/StoreMain';
 
+
 export default function Store() {
+  
+
   return (
     <>
       <StoreMain />
     </>
   );
 }
+
